Decoupled auth form from the Homepage.
